Make CircularArcProperties public
